Integration: Breakup resiliency app tests
Break up resiliency app tests into separate files for better
organization and ensure that we don't hit timeouts in slow CI runners.
